Fixed bugs in UnorderedList.pop()
The major problem was that I forgot that 0 is a falsey value that
therefore did not pass the checks. Solved this by first checking
that the argument *position* is has a proper index and then that
it is an integer.

Also improved tests to get 98% code-coverage. To ease debugging and
make the UnorderedList-class useful, I also added a helper-method
that returns a copy of all present values in the list.