Implement BST traverse() and __str__() methods.
Includes tests with good coverage and linted code.

Also abstracted away the creation of BSTs with a helper method
that will create new trees from a given sequence of keys.
Unfortunately it does not accept dictionaries with key:value pairs
but for the simple testing done so far in the project, it's enough
if the helper can create trees with just keys.